provider/vagrant: adapt to new API
